Split in 2 commands: phcp-{prepare,run}-interfov-registration
